temeprature bitshifting fix

+++ Core/Inc/adbms_to_CAN.h
@@ -24,7 +24,7 @@ HAL_StatusTypeDef ADBMS_CAN_SendVoltages_Module(CAN_HandleTypeDef *hcan, uint8_t
 HAL_StatusTypeDef ADBMS_CAN_SendTemperatures_Module(CAN_HandleTypeDef *hcan, uint8_t module);
 HAL_StatusTypeDef ADBMS_CAN_SendAll(CAN_HandleTypeDef *hcan, AMSStates_t ams_current_state);
 
-uint16_t data_to_volts(int16_t, adbms_data_type_t);
+float data_to_volts(int16_t, adbms_data_type_t);
 float getTemperatureCAN(int16_t);
 
 #endif /* INC_ADBMS_TO_CAN_H_ */
